fix: move topology details to separate components
