+++ frontend/packages/rhoas-plugin/locales/en/rhoas-plugin.json
@@ -5,18 +5,25 @@
   "Unlock with token": "Unlock with token",
   "Unlocked": "Unlocked",
   "RHOAS can include Streams for Kafka, Service Registry": "RHOAS can include Kafka, Service Registry and other Red Hat Services",
+  "KafkaCardDescription": "OpenShift Streams for Apache Kafka",
+  "ManagedServices-card-description": "Red Hat OpenShift Application Services provide a streamlined developer experience for building, deploying, and scaling cloud-native applications in open hybrid-cloud environments. The combination of enterprise-grade Kubernetes, cloud-native approach toapplication delivery, and managed operations allows teams to focus on core competencies, accelerate time to value and reduce operational cost.",
   "Access Red Hat Cloud Services with API Token": "Access Red Hat Cloud Services with API Token",
-  "To access this Cloud Service, input the API token which can be located at": "To access this Cloud Service, input the API token which can be located at ",
-  "API token can be accessed at": "API token can be accessed at",
+  "Problem with creating secret": "Problem with creating secret: {{error}} Namespace: {{namespace}}",
+  "Cannot create service account": "Cannot create service account: {{error}} Namespace: {{namespace}}",
+  "To access this Cloud Service, input the API token which can be located at <2>https://cloud.redhat.com/openshift/token</2>": "To access this Cloud Service, input the API token which can be located at <2>https://cloud.redhat.com/openshift/token</2>",
+  "API Token": "API Token",
   "Cant create an access token? Contact your administrator": "Cant create an access token? Contact your administrator",
   "Create": "Create",
-  "Cancel": "Cancel",
   "Reset": "Reset",
   "Select Kafka Cluster": "Select Kafka Cluster",
   "The Kafka cluster selected below will appear on the topology view.": "The Kafka cluster selected below will appear on the topology view.",
   "All Kafka clusters are in use": "All Kafka clusters are in use",
   "Go back to Services Catalog": "Go back to Services Catalog",
   "No Kafka Clusters found": "No Kafka Clusters found",
+  "Cancel": "Cancel",
+  "Could not fetch services": "Could not fetch services",
+  "Could not connect to RHOAS with API Token": "Could not connect to RHOAS with API Token",
+  "Failed to load list of services": "Failed to load list of services: {{error}}",
   "Search by name": "Search by name",
   "Search by name...": "Search by name...",
   "Cluster Name": "Name",
@@ -28,15 +35,7 @@
   "List of Kafka Instances": "List of Kafka Instances",
   "Bootstrap Server": "Bootstrap Server",
   "URL": "URL",
-  "Failed to load list of services": "Failed to load list of services: {{error}}",
-  "Could not fetch services": "Could not fetch services",
-  "Could not connect to RHOAS with API Token": "Could not connect to RHOAS with API Token",
   "Cloud Service": "Cloud Service",
   "This resource represents service that exist outside your cluster": "This resource represents service that exist outside your cluster",
-  "Create a binding connector": "Create a binding connector",
-  "API Token": "API Token",
-  "Problem with creating secret": "Problem with creating secret: {{error}} Namespace: {{namespace}}",
-  "Cannot create service account": "Cannot create service account: {{error}} Namespace: {{namespace}}",
-  "KafkaCardDescription": "OpenShift Streams for Apache Kafka",
-  "ManagedServices-card-description": "Red Hat OpenShift Application Services provide a streamlined developer experience for building, deploying, and scaling cloud-native applications in open hybrid-cloud environments. The combination of enterprise-grade Kubernetes, cloud-native approach toapplication delivery, and managed operations allows teams to focus on core competencies, accelerate time to value and reduce operational cost."
+  "Create a binding connector": "Create a binding connector"
 }
