+++ frontend/packages/rhoas-plugin/locales/en/topology.json
