+++ frontend/public/locales/en/public.json
@@ -731,9 +731,6 @@
   "Enabled": "Enabled",
   "Failed to create connection": "Failed to create connection",
   "Please try again": "Please try again",
-  "Could not fetch services": "Could not fetch services",
-  "Could not connect to RHOAS with API Token": "Could not connect to RHOAS with API Token",
-  "Failed to load list of services": "Failed to load list of services",
   "No results found": "No results found",
   "No results match the filter criteria. Remove all filters or clear all filters to show results.": "No results match the filter criteria. Remove all filters or clear all filters to show results."
 }
